Gate account routes from inside Suspense to avoid PPR bailouts
When auth is off, layout-level `notFound()` in /account/(authenticated)
and /account/login was triggering useSearchParams() bailouts in Next 16
canary's PPR prerenderer (the throw escapes before any Suspense
boundary, hitting Next's internal not-found path which is missing a
Suspense). The build still succeeded but no static HTML was emitted for
/account/*.

Move the `isAuthEnabled` check into server components that are already
inside Suspense boundaries:

- (authenticated)/layout.tsx — gate in AccountGate and UserInfo (both
  rendered under Suspense)
- login/layout.tsx — slimmed to `<Suspense>{children}</Suspense>`;
  gate moved to login/page.tsx (server)
- login/page.tsx — server component that gates and renders the new
  <LoginRedirect /> client subcomponent
- login/login-redirect.tsx — extracted client redirect logic

Result: all six account routes prerender to HTML, 0 bailouts.
