+++ apps/template/app/account/login/layout.tsx
@@ -1,19 +1,5 @@
-import type { Metadata } from "next";
-import { getTranslations } from "next-intl/server";
-import { notFound } from "next/navigation";
-
-import { isAuthEnabled } from "@/lib/auth";
-
-export async function generateMetadata(): Promise<Metadata> {
-  const t = await getTranslations("seo");
-  return {
-    title: t("loginTitle"),
-    robots: { index: false, follow: false },
-  };
-}
+import { Suspense } from "react";
 
 export default function LoginLayout({ children }: { children: React.ReactNode }) {
-  if (!isAuthEnabled) notFound();
-
-  return children;
+  return <Suspense>{children}</Suspense>;
 }
